tests/wpt/tests/css/css-typed-om/the-stylepropertymap/properties/offset-rotate.html

<meta charset="utf-8">
<title>'offset-rotate' property</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

runPropertyTests('offset-rotate', [
  {
    syntax: 'auto',
    computed: (input, result) => {
      try {
        assert_is_unsupported(result);
      } catch(e) {
        assert_style_value_equals(result, input);
      }
    }
  },
  {
    syntax: 'reverse',
    computed: (_, result) => assert_is_unsupported(result)
  },
  {
    syntax: '<angle>',
    specified: (input, result, alternateExpected) => {
      try {
        assert_style_value_equals(result, input);
      } catch(e) {
        if (alternateExpected === undefined) {
          throw e;
        }
        assert_style_value_equals(result, alternateExpected);
      }
    },
    computed: (input, result) => {
      const degrees = input.to('deg');
      assert_style_value_equals(result, degrees);
    }
  },
]);

runUnsupportedPropertyTests('offset-rotate', [
  'auto 90deg',
  'reverse -90deg',
]);

</script>
