third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/offset-rotate-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'offset-rotate' to an angle: calc(0rad + 0deg)
  assert_equals: expected 1 but got 2
Harness: the test ran to completion.
